+++ apps/frontend/src/modules/points-of-interest/libs/enums/enums.ts
@@ -1,1 +1,2 @@
+export { PointOfInterestNotification } from "./point-of-interest-notification.enum.js";
 export { PointsOfInterestApiPath } from "@smartscapes/shared";