+++ apps/frontend/src/modules/routes/slices/routes.slice.ts
@@ -11,26 +11,22 @@ import {
 	create,
 	discardCreateRouteFormData,
 	getAll,
-	getRouteById,
-	patchRoute,
 	restoreCreateRouteFormData,
 } from "./actions.js";
 
 type State = {
+	createRouteFormData: null | Partial<RouteCreateRequestDto>;
 	createStatus: ValueOf<typeof DataStatus>;
 	dataStatus: ValueOf<typeof DataStatus>;
-	formData: null | Partial<RouteCreateRequestDto>;
-	formDataStatus: ValueOf<typeof DataStatus>;
-	route: null | RouteGetByIdResponseDto;
+	restoreCreateRouteFormStatus: ValueOf<typeof DataStatus>;
 	routes: RouteGetByIdResponseDto[];
 };
 
 const initialState: State = {
+	createRouteFormData: null,
 	createStatus: DataStatus.IDLE,
 	dataStatus: DataStatus.IDLE,
-	formData: null,
-	formDataStatus: DataStatus.IDLE,
-	route: null,
+	restoreCreateRouteFormStatus: DataStatus.IDLE,
 	routes: [],
 };
 
@@ -47,17 +43,6 @@ const { actions, name, reducer } = createSlice({
 			state.createStatus = DataStatus.REJECTED;
 		});
 
-		builder.addCase(getRouteById.pending, (state) => {
-			state.dataStatus = DataStatus.PENDING;
-		});
-		builder.addCase(getRouteById.fulfilled, (state, action) => {
-			state.route = action.payload.data;
-			state.dataStatus = DataStatus.FULFILLED;
-		});
-		builder.addCase(getRouteById.rejected, (state) => {
-			state.dataStatus = DataStatus.REJECTED;
-		});
-
 		builder.addCase(getAll.pending, (state) => {
 			state.dataStatus = DataStatus.PENDING;
 		});
@@ -68,30 +53,19 @@ const { actions, name, reducer } = createSlice({
 			state.dataStatus = DataStatus.REJECTED;
 		});
 
-		builder.addCase(patchRoute.pending, (state) => {
-			state.dataStatus = DataStatus.PENDING;
-		});
-		builder.addCase(patchRoute.fulfilled, (state, action) => {
-			state.route = action.payload.data;
-			state.dataStatus = DataStatus.FULFILLED;
-		});
-		builder.addCase(patchRoute.rejected, (state) => {
-			state.dataStatus = DataStatus.REJECTED;
-		});
-
 		builder.addCase(restoreCreateRouteFormData.pending, (state) => {
-			state.formDataStatus = DataStatus.PENDING;
+			state.restoreCreateRouteFormStatus = DataStatus.PENDING;
 		});
 		builder.addCase(restoreCreateRouteFormData.fulfilled, (state, action) => {
-			state.formData = action.payload;
-			state.formDataStatus = DataStatus.FULFILLED;
+			state.createRouteFormData = action.payload;
+			state.restoreCreateRouteFormStatus = DataStatus.FULFILLED;
 		});
 		builder.addCase(restoreCreateRouteFormData.rejected, (state) => {
-			state.formDataStatus = DataStatus.REJECTED;
+			state.restoreCreateRouteFormStatus = DataStatus.REJECTED;
 		});
 
 		builder.addCase(discardCreateRouteFormData.fulfilled, (state) => {
-			state.formData = null;
+			state.createRouteFormData = null;
 		});
 	},
 	initialState,