+++ apps/frontend/src/modules/routes/slices/routes.ts
@@ -2,23 +2,26 @@ import {
 	create,
 	discardCreateRouteFormData,
 	getAll,
-	getRouteById,
-	patchRoute,
+	getById,
+	patch,
 	preserveCreateRouteFormData,
 	restoreCreateRouteFormData,
 } from "./actions.js";
-import { actions } from "./routes.slice.js";
+import { actions as routeDetailsActions } from "./route-details.slice.js";
+import { actions as routesActions } from "./routes.slice.js";
 
 const allActions = {
-	...actions,
+	...routesActions,
+	...routeDetailsActions,
 	create,
 	discardCreateRouteFormData,
 	getAll,
-	getRouteById,
-	patchRoute,
+	getById,
+	patch,
 	preserveCreateRouteFormData,
 	restoreCreateRouteFormData,
 };
 
 export { allActions as actions };
-export { reducer } from "./routes.slice.js";
+export { reducer as routeDetailsReducer } from "./route-details.slice.js";
+export { reducer as routesReducer } from "./routes.slice.js";