+++ packages/shared/src/modules/points-of-interest/libs/validation-schemas/points-of-interest-create.validation-schema.ts
@@ -13,6 +13,7 @@ import {
 const pointsOfInterestCreate = z.strictObject({
 	description: z
 		.string()
+		.trim()
 		.max(PointsOfInterestValidationRule.DESCRIPTION_MAX_LENGTH, {
 			message: PointsOfInterestValidationMessage.DESCRIPTION_MAXIMUM_LENGTH,
 		})