+++ packages/shared/src/modules/user-routes/validation-schemas/user-route-patch.validation-schema.ts
@@ -8,40 +8,39 @@ import {
 import { UserRouteValidationMessage } from "../libs/enums/enums.js";
 
 const userRoutePatch = z.strictObject({
-	actualGeometry: z.object(
-		{
-			coordinates: z.tuple([
-				z
-					.number()
-					.min(
-						CoordinatesValidationRule.LONGITUDE_MIN,
-						CoordinatesValidationMessage.LONGITUDE_MIN,
-					)
-					.max(
-						CoordinatesValidationRule.LONGITUDE_MAX,
-						CoordinatesValidationMessage.LONGITUDE_MAX,
-					),
-				z
-					.number()
-					.min(
-						CoordinatesValidationRule.LATITUDE_MIN,
-						CoordinatesValidationMessage.LATITUDE_MIN,
-					)
-					.max(
-						CoordinatesValidationRule.LATITUDE_MAX,
-						CoordinatesValidationMessage.LATITUDE_MAX,
-					),
-			]),
-			type: z.literal(LocationType.LINE_STRING, {
-				message: UserRouteValidationMessage.INVALID_LOCATION_TYPE,
+	actualGeometry: z.object({
+		coordinates: z
+			.array(
+				z.tuple([
+					z
+						.number()
+						.min(
+							CoordinatesValidationRule.LONGITUDE_MIN,
+							CoordinatesValidationMessage.LONGITUDE_MIN,
+						)
+						.max(
+							CoordinatesValidationRule.LONGITUDE_MAX,
+							CoordinatesValidationMessage.LONGITUDE_MAX,
+						),
+					z
+						.number()
+						.min(
+							CoordinatesValidationRule.LATITUDE_MIN,
+							CoordinatesValidationMessage.LATITUDE_MIN,
+						)
+						.max(
+							CoordinatesValidationRule.LATITUDE_MAX,
+							CoordinatesValidationMessage.LATITUDE_MAX,
+						),
+				]),
+			)
+			.min(CoordinatesValidationRule.MIN_COORDINATES_COUNT, {
+				message: CoordinatesValidationMessage.INVALID_COORDINATES,
 			}),
-		},
-		{
-			error: () => {
-				return UserRouteValidationMessage.ACTUAL_GEOMETRY_REQUIRED;
-			},
-		},
-	),
+		type: z.literal(LocationType.LINE_STRING, {
+			message: UserRouteValidationMessage.INVALID_LOCATION_TYPE,
+		}),
+	}),
 	routeId: z.number({
 		message: UserRouteValidationMessage.ROUTE_ID_INVALID_TYPE,
 	}),